do not attempt to index an event that does not have "Indexed" interface.
it is possible, though not likely, that an event returned from the event store
has not been properly filtered as only an 'indexable' event during reindexing.
to address this I have built a wall, just kidding, I check to see if it's got the
"Indexed" interface before indexing.

take 2, wtf.  if NOT!! indexed