modifies the pbj_template controller helper method and twig extension to return a symfony best practices template location.  http://symfony.com/doc/current/best_practices/templates.html#template-locations
This also allows for the traditional bundle + override style development where a bundle provides a base
implementation (that you'd point to with your twig namespaced path) and then your app/bundle can
substitute it.

The primary reason that pbj resolution to template locations does not use standard bundle conventions is
that pbj schemas are independent of bundle and should have no knowledge that they might have templates
that originate from a bundle.  They must remain agnostic to any framework in order to be reusable
in and outside of symfony.  The other big benefit to having this method is to prevent having to do switch
statements inside of twig to figure out what template to use for the given pbj.