Initial release (#6)
* added check health handler

* upated check health handler log to include class name.  makes message more unique and easier to grep

* upated check health handler log to include class name.  makes message more unique and easier to grep

* upated check health handler log to include class name.  makes message more unique and easier to grep

* for the event search, use 500ms as query_timeout default

* moved shard utils and status code converter to pbjx

* add handler guesser

* added export commands for event store

* add guesser create handler (wip)

* formatting

* wip

* allow for skipping of invalid json line OR general exceptions

* php7.1 updates

* wip

* wip

* the php7ing

* defaults in wrong place

* guesser is not running, so logger aware trait is not used here, wire it explicitly

* provider is optional

* dash convention for worker ids and by default, process requests too

* ctx_request now set by default in pbjx process

* add defaults for config

* php7.1 updates and added convenience method to loading schemas by mixin in commands

* the php7ing and begin cleanup to use latest updates from pbjx lib

* Updated all Form files to php7

* Revert php7 upgrade for Symfony extended methods

* Renamed Pbjx classes ("Default" to "Simple")

* Added missing ";"

* the php7ing - command cleanup - use "pipe" methods and create generate event store storage creator (wip)

* RE #4: Fixed KeyValueType

Also added test unit

* php7ing cleanup - add generic command to creat event store and search and make replay/reindex/export all one command and make streamid optional

* add tenant_id_field config on event search which can be used to route indexing based on message content/context

* add tenant_id_field config on event search which can be used to route indexing based on message content/context (set into container as well)

* add query parsing binder for convenience

* remove double print of context

* allow for empty payload.. sometimes the message curie itself is all you need to process the message

* the php7ing - cdo visibility and allow null on receive key (validates it later anyways)

* ensure json_decode is given a string when processing json from request

* cast tenant_id to string in all commands

* fixup "ReindexEventsCommand". dynamodb parallel processing with no range key creates some
inconsistent chunk sizes but the reindexer is now making a copy of the array once it reaches the
expected size so the resulting batch sizes are still as expected.  removed output to eliminate
confusion with event store batching which may be different from the batching sent for reindexing.

* Fixed calling "collection_item_prototype" macro in collection widget

* do not attempt to index an event that does not have "Indexed" interface.

it is possible, though not likely, that an event returned from the event store
has not been properly filtered as only an 'indexable' event during reindexing.
to address this I have built a wall, just kidding, I check to see if it's got the
"Indexed" interface before indexing.

* do not attempt to index an event that does not have "Indexed" interface.

it is possible, though not likely, that an event returned from the event store
has not been properly filtered as only an 'indexable' event during reindexing.
to address this I have built a wall, just kidding, I check to see if it's got the
"Indexed" interface before indexing.

take 2, wtf.  if NOT!! indexed

* add methods to render pbj and pbj with a form.  eliminates repetitive code in controllers

* Fixed mapping data to form for special field types

* adding basic documentation for pbjx configuration and use.

* add examples of using the pbjx http endpoints

* add examples of using the pbjx http endpoints - fix typo

* added compiler pass that can automatically alias handlers for concrete handlers to allow library work against mixins that just work when installed

* adding documentation and a contrived example of how to use mixins to develop libraries
and how this bundle can help resolve service references dynamically when providing
implementations based on mixins.

* updated php docblock on class to reference readme which has more details on how to use an "Aliased" handler.

* when listeners are registered we should resolve the value so we can use container parameters in event listener
event names.  what the shell?  super cool inorite, by using the %app_vendor% parameter (or any other) a bundle
can provide a listener which listens to your events.  this makes pbjx bindings much simpler and requires
a lot less configuration.

* modifies the pbj_template controller helper method and twig extension to return a symfony best practices
template location.  http://symfony.com/doc/current/best_practices/templates.html#template-locations

This also allows for the traditional bundle + override style development where a bundle provides a base
implementation (that you'd point to with your twig namespaced path) and then your app/bundle can
substitute it.

The primary reason that pbj resolution to template locations does not use standard bundle conventions is
that pbj schemas are independent of bundle and should have no knowledge that they might have templates
that originate from a bundle.  They must remain agnostic to any framework in order to be reusable
in and outside of symfony.  The other big benefit to having this method is to prevent having to do switch
statements inside of twig to figure out what template to use for the given pbj.

* organize config a little bit (cdo)

* prep for initial release.  use pbjx ~2.0 constraint